feat: 🎉 implement Superblock class
Created a class to handle parsing the superblock bytes from the volume.
It takes a ByteBuffer in the constructor and reads the relevant values
(in relation to the coursework goals, reading files from the volume),
storing them in members of the object. The values are accessible to
other objects through getter methods.