refactor: 💡 use Superblock object in Volume
I've rewritten most of the code in the Volume class to make use of the
new Superblock class, removing as much "boilerplate" as possible. The
initialisation logic has been moved into the constructor which creates a
bytebuffer from bytes 1024-2047 and constructs a new Superblock object.

User facing properties of the volume, such as it's label and capacity,
are made available with getter methods.

These properties can be printed to the console for debugging with the
Volume#printDebugInfo() method.